+++ incs/minishell.h
@@ -6,7 +6,7 @@
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/25 10:08:39 by marvin            #+#    #+#             */
-/*   Updated: 2023/09/23 10:48:11 by codespace        ###   ########.fr       */
+/*   Updated: 2023/09/23 11:59:48 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,18 +38,12 @@
 # define MYNAME "minishell"
 
 /* error name macros */
-# define ERR_PIPE "pipe"
-# define ERR_FORK "fork"
-# define ERR_DUP "dup"
-# define ERR_OPEN "open"
-# define ERR_DUP2 "dup2"
-# define ERR_MALLOC "malloc"
 # define ERR_CMD "command not found"
 
 /* error code macros */
-
 # define CODE_OPEN 1
 # define CODE_CMD 127
+# define CODE_DUP 123123     //i don't know...
 # define CODE_DUP2 123123     //i don't know...
 # define CODE_EXECVE 123123     //i don't know...
 
@@ -142,9 +136,10 @@ enum e_builtin
 	BI_UNSET,
 	BI_EXPORT,
 };
-
+//////////////////////////////////////
 //////////////////////////////////////
 ////////////// STRUCTS ///////////////
+//////////////////////////////////////
 //////////////////////////////////////
 
 /*ms_setup.c*/
@@ -159,24 +154,34 @@ t_block	*block_init(t_ms *ms, t_block *manager, t_token_list *prompt, int my_id)
 void	block_destroy(void *og_block);
 
 /* ms_signals.c */
-
-int		save_signal(int *num);
-int		check_for_signals(t_ms *ms);
 void	signal_handler(int signum);
 int		ms_prepare_signal(t_ms *ms, void (*handler)(int));
-t_ms	*sigint_heredoc_where_ms_is(t_ms *ms);
-void	signal_handler_exec(int signum);
 int		ms_reset_signal(t_ms *ms);
-void	signal_waiting(int signum);
-int		ms_signal_replace(t_ms *ms, void (*handler)(int));
-/* ms_prompt.c */
-
-
-
-//functions to split prompt into children
 
 //////////////////////////////////////
+//////////////////////////////////////
+//////////////// LEXER ///////////////
+//////////////////////////////////////
+//////////////////////////////////////
+
+/*get_prompt.c*/
+int				get_prompt(t_ms *ms);
+char			*prompt_readline(t_ms *ms);
+int				prompt_token_setup(t_ms *ms, char *line);
+
+
+/* ??????? */																	// discriminar por ficheiros
+int				ft_isquote(int c);
+char			*ft_strdup_len(char *s, int len);
+int				prompt_is_valid(t_ms *ms);
+int				valid_elem_order(t_ms *ms);
+int				valid_redir_texts(t_ms *ms);
+int				invalid_elem_msg(t_ms *ms, int type, char *has_text, char *text);
+
+//////////////////////////////////////
+//////////////////////////////////////
 ////////////// PARSER ////////////////
+//////////////////////////////////////
 //////////////////////////////////////
 
 /* task_distributor.c*/
@@ -196,7 +201,6 @@ void	manager_destroy_worker_tasks(t_block *manager);
 int		worker_task_preparation(t_block *worker);
 int		worker_extract_redirections(t_block *worker);
 
-
 /*worker_tasks_2.c*/
 int 	worker_args_split_add_token(t_block *worker, t_token_node *arg, int *move);
 int		worker_args_rm_unguarded_quotes(t_block *worker);
@@ -204,9 +208,10 @@ int		worker_args_split_unguarded_quotes(t_block *worker);
 int		worker_args_expand_dollar_wildcard(t_block *worker);
 int		worker_dump_tasks_to_cmd_args(t_block *worker);
 
-
 //////////////////////////////////////
-////////////// EXECUTER ////////////////
+//////////////////////////////////////
+////////////// EXECUTER //////////////
+//////////////////////////////////////
 //////////////////////////////////////
 
 /*execution_tree.c*/
@@ -229,15 +234,21 @@ int		setup_execution_tree(t_ms *ms, t_block *manager, t_token_list *prompt, int 
 int		exec_command(t_block *block);
 int		process_execution(t_block *block);
 
+/*child_process.c*/
+int		child_process(t_block *block);
+void	error_child_exit(t_block *block, char *error_msg, int errcode,
+		int with_ms);
+void	perror_child_exit(t_block *block, int errcode,
+		int with_ms);
 
-
-void	error_child_exit(t_block *block, char *function, char *cmd, int errcode, int with_ms);
+/*child_execve.c*/
+int		exec_command(t_block *block);
 int		perror_msg_func(t_block *block, char *function, int errcode, int with_ms);
-void	perror_child_exit(t_block *block, char *function, int errcode, int with_ms);
-
 
 //////////////////////////////////////
+//////////////////////////////////////
 //////////// REDIRECTIONS ////////////
+//////////////////////////////////////
 //////////////////////////////////////
 
 /*prepare_redirections.c*/
@@ -249,7 +260,6 @@ int		manage_io_expansion(t_block *block);
 /*close_in_out.c*/
 void	close_in_fds(t_block *block);
 void	close_out_fds(t_block *block);
-
 
 /*heredoc_open*/
 int		open_here_docs_at_block(t_block *block);
@@ -263,9 +273,10 @@ int		here_doc(t_block *block, char *eof, int has_quote_guard);
 /*heredoc_open.c*/
 int		open_here_docs_at_block(t_block *block);
 
-
+//////////////////////////////////////
 //////////////////////////////////////
 //////////// EXPANSIONS //////////////
+//////////////////////////////////////
 //////////////////////////////////////
 
 /*wildcard_expansion.c*/
@@ -292,14 +303,17 @@ int		here_doc_expand_dollars(char **to_expand, t_ms *ms);
 /*rm_unguarded_quotes*/
 int		remove_unguarded_quotes(char **str, int *has_guards);
 
-
+//////////////////////////////////////
 //////////////////////////////////////
 //////////// BUILTINS ////////////////
 //////////////////////////////////////
+//////////////////////////////////////
 
+/*bi_check_and_exec.c*/
 int		check_builtins(t_block *block);
 int		exec_builtin(t_block *block, int builtin);
 
+/* ??????? */																// discriminar por ficheiros
 int		run_cd(t_block *block);
 int		run_env(t_block *block);
 int		run_pwd(t_block *block);
@@ -310,44 +324,16 @@ int		run_exit(t_block *block);
 
 int		set_beg_path(t_block *block, char **curpath, char *curr);
 int		cd_exists(t_block *block);
-
-int		set_beg_path(t_block *block, char **curpath, char *curr);
-int		cd_exists(t_block *block);
-
-char	*ft_strchr(const char *s, int c);
 int		get_corr_env(t_block *block, char *arg, int is_exporting);
 int		env_add(t_block *block, char *new);
-
-int		ft_isdigit(int c);
-int		ft_isalpha(int c);
-
 int		env_remove(t_block *block, int index);
 
-
-
-
 //////////////////////////////////////
-//////////////// LEXER ///////////////
-//////////////////////////////////////
-
-int				get_prompt(t_ms *ms);
-int				setup_prompt(t_ms *ms);
-int				prompt_token_setup(t_ms *ms, char *line);
-char			*prompt_readline(t_ms *ms);
-
-int				ft_isquote(int c);
-char			*ft_strdup_len(char *s, int len);
-
-int				prompt_is_valid(t_ms *ms);
-int				valid_elem_order(t_ms *ms);
-int				valid_redir_texts(t_ms *ms);
-int				invalid_elem_msg(t_ms *ms, int type, char *has_text, char *text);
-
 //////////////////////////////////////
 //////////// GENERIC UTILS ///////////
 //////////////////////////////////////
+//////////////////////////////////////
 
-/* future_libft.c*/
 int		ft_matrixlen(void *mat);
 int		ft_charmatdup(char ***dest, char **src);
 void	*quicksort_pointers(void *arr, int size, int (*cmp)(void *, void *));
