+++ srcs/lexer/get_prompt.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ms_prompt.c                                        :+:      :+:    :+:   */
+/*   get_prompt.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/17 00:13:38 by mmaria-d          #+#    #+#             */
-/*   Updated: 2023/09/23 10:50:16 by codespace        ###   ########.fr       */
+/*   Updated: 2023/09/23 11:57:02 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
