+++ test1
@@ -1,0 +1,8 @@
+Makefile
+incs
+libft
+minishell
+objs
+srcs
+test1
+todolist.txt
